# api/gen/http/admin/v1/admin.gen.go
//go:build go1.22

// Package adminv1http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package adminv1http

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// PurgeTagsRequest Request to purge cache entries by surrogate-key tags.
type PurgeTagsRequest struct {
	// Tags Surrogate-key tags to purge.
	Tags []string `json:"tags"`
}

// PurgeURLRequest Request to purge a single cached response by URL.
type PurgeURLRequest struct {
	// Method HTTP method of the cached response to purge.
	Method *string `json:"method,omitempty"`

	// Url URL path to purge (e.g. "/v1/menus/menu-demo/published").
	Url string `json:"url"`
}

// PurgeResponseCacheURLJSONRequestBody defines body for PurgeResponseCacheURL for application/json ContentType.
type PurgeResponseCacheURLJSONRequestBody = PurgeURLRequest

// PurgeAuthCacheTagsJSONRequestBody defines body for PurgeAuthCacheTags for application/json ContentType.
type PurgeAuthCacheTagsJSONRequestBody = PurgeTagsRequest

// PurgeResponseCacheTagsJSONRequestBody defines body for PurgeResponseCacheTags for application/json ContentType.
type PurgeResponseCacheTagsJSONRequestBody = PurgeTagsRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Purge a single cached response by URL.
	// (POST /v1/cache/purge)
	PurgeResponseCacheURL(w http.ResponseWriter, r *http.Request)
	// Purge auth cache entries by surrogate-key tags.
	// (POST /v1/cache/purge/auth)
	PurgeAuthCacheTags(w http.ResponseWriter, r *http.Request)
	// Purge response cache entries by surrogate-key tags.
	// (POST /v1/cache/purge/tags)
	PurgeResponseCacheTags(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PurgeResponseCacheURL operation middleware
func (siw *ServerInterfaceWrapper) PurgeResponseCacheURL(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PurgeResponseCacheURL(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PurgeAuthCacheTags operation middleware
func (siw *ServerInterfaceWrapper) PurgeAuthCacheTags(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PurgeAuthCacheTags(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PurgeResponseCacheTags operation middleware
func (siw *ServerInterfaceWrapper) PurgeResponseCacheTags(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PurgeResponseCacheTags(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("POST "+options.BaseURL+"/v1/cache/purge", wrapper.PurgeResponseCacheURL)
	m.HandleFunc("POST "+options.BaseURL+"/v1/cache/purge/auth", wrapper.PurgeAuthCacheTags)
	m.HandleFunc("POST "+options.BaseURL+"/v1/cache/purge/tags", wrapper.PurgeResponseCacheTags)

	return m
}

type PurgeResponseCacheURLRequestObject struct {
	Body *PurgeResponseCacheURLJSONRequestBody
}

type PurgeResponseCacheURLResponseObject interface {
	VisitPurgeResponseCacheURLResponse(w http.ResponseWriter) error
}

type PurgeResponseCacheURL204Response struct {
}

func (response PurgeResponseCacheURL204Response) VisitPurgeResponseCacheURLResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type PurgeResponseCacheURL404Response struct {
}

func (response PurgeResponseCacheURL404Response) VisitPurgeResponseCacheURLResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type PurgeAuthCacheTagsRequestObject struct {
	Body *PurgeAuthCacheTagsJSONRequestBody
}

type PurgeAuthCacheTagsResponseObject interface {
	VisitPurgeAuthCacheTagsResponse(w http.ResponseWriter) error
}

type PurgeAuthCacheTags204Response struct {
}

func (response PurgeAuthCacheTags204Response) VisitPurgeAuthCacheTagsResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type PurgeResponseCacheTagsRequestObject struct {
	Body *PurgeResponseCacheTagsJSONRequestBody
}

type PurgeResponseCacheTagsResponseObject interface {
	VisitPurgeResponseCacheTagsResponse(w http.ResponseWriter) error
}

type PurgeResponseCacheTags204Response struct {
}

func (response PurgeResponseCacheTags204Response) VisitPurgeResponseCacheTagsResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Purge a single cached response by URL.
	// (POST /v1/cache/purge)
	PurgeResponseCacheURL(ctx context.Context, request PurgeResponseCacheURLRequestObject) (PurgeResponseCacheURLResponseObject, error)
	// Purge auth cache entries by surrogate-key tags.
	// (POST /v1/cache/purge/auth)
	PurgeAuthCacheTags(ctx context.Context, request PurgeAuthCacheTagsRequestObject) (PurgeAuthCacheTagsResponseObject, error)
	// Purge response cache entries by surrogate-key tags.
	// (POST /v1/cache/purge/tags)
	PurgeResponseCacheTags(ctx context.Context, request PurgeResponseCacheTagsRequestObject) (PurgeResponseCacheTagsResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// PurgeResponseCacheURL operation middleware
func (sh *strictHandler) PurgeResponseCacheURL(w http.ResponseWriter, r *http.Request) {
	var request PurgeResponseCacheURLRequestObject

	var body PurgeResponseCacheURLJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PurgeResponseCacheURL(ctx, request.(PurgeResponseCacheURLRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PurgeResponseCacheURL")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PurgeResponseCacheURLResponseObject); ok {
		if err := validResponse.VisitPurgeResponseCacheURLResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PurgeAuthCacheTags operation middleware
func (sh *strictHandler) PurgeAuthCacheTags(w http.ResponseWriter, r *http.Request) {
	var request PurgeAuthCacheTagsRequestObject

	var body PurgeAuthCacheTagsJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PurgeAuthCacheTags(ctx, request.(PurgeAuthCacheTagsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PurgeAuthCacheTags")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PurgeAuthCacheTagsResponseObject); ok {
		if err := validResponse.VisitPurgeAuthCacheTagsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PurgeResponseCacheTags operation middleware
func (sh *strictHandler) PurgeResponseCacheTags(w http.ResponseWriter, r *http.Request) {
	var request PurgeResponseCacheTagsRequestObject

	var body PurgeResponseCacheTagsJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PurgeResponseCacheTags(ctx, request.(PurgeResponseCacheTagsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PurgeResponseCacheTags")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PurgeResponseCacheTagsResponseObject); ok {
		if err := validResponse.VisitPurgeResponseCacheTagsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}